Update audio-guestbook.ino
Includes a new default greeting option, a way of recording a custom greeting by pressing the playback button twice with 200 + 600 milli seconds when the phone is in prompting mode. You can also delete the custom greeting by pressing the playback button twice with 200 + 600 milli seconds when the phone is in Ready mode (handset is in place).

You need to copy the default_greeting.wav and greeting_record_prompt.wav to the root of the SD card.